Add files via upload
This project involves querying a database to retrieve various pieces of information related to offices, orders, employees, customers, payments, and products. It starts with simple queries like showing all information in the offices relation or extracting specific columns like city and phone of offices. Then, it progresses to more complex queries such as filtering data based on certain conditions, calculating averages, counts, and identifying patterns within the data. For instance, it identifies large orderlines, doubles quantity In Stock for specific products, and determines the mean number of tosses required for certain conditions in a simulated experiment. Additionally, it extracts data based on specific criteria like orders with a status of 'Disputed' or products that have never been ordered. Furthermore, it involves aggregating data to show statistics like the number of employees in each office or the number of customers associated with each employee meeting specific conditions. Lastly, it includes more advanced queries like identifying pairs of offices in the same country or checking for any attribute storing a price that contains a zero value. Through these queries, the project aims to extract meaningful insights from the given database and answer various analytical questions related to different aspects of the business operations.